only Author access to all the entites
Updated handlers and repositories of each entity: notebook, unit, page, note.

Added Aouthor check for each.
Corrected counting.

+++ Application/Notebooks/Delete.cs
@@ -1,5 +1,6 @@
 ﻿using Application.Core;
 using Domain.Repositories.Repos.Interfaces;
+using IndentityLogic.Interfaces;
 using MediatR;
 
 namespace Application.Notebooks
@@ -14,10 +15,12 @@ namespace Application.Notebooks
         public class Handler : IRequestHandler<Command, Result<MediatR.Unit>>
         {
             private readonly INotebookRepository _notebookRepository;
+            private readonly IUserAccessor _userAccessor;
 
-            public Handler(INotebookRepository notebookRepository)
+            public Handler(INotebookRepository notebookRepository, IUserAccessor userAccessor)
             {
                 _notebookRepository = notebookRepository;
+                _userAccessor = userAccessor;
             }
 
             public async Task<Result<MediatR.Unit>> Handle(Command request, CancellationToken cancellationToken)
@@ -25,6 +28,12 @@ namespace Application.Notebooks
                 var notebook = await _notebookRepository.GetOneAsync(request.Id);
 
                 if (notebook == null) return null;
+
+                if (!await _notebookRepository.IfUserHasAccessToTheNotebook(request.Id,
+                    _userAccessor.GetUserId()))
+                {
+                    return Result<MediatR.Unit>.Failure("You have no access to this data");
+                }
 
                 var result = await _notebookRepository.RemoveAsync(notebook) > 0;
